refactor: create AbstractEventHandler to remove duplication in persistence logic
